nutrition-analyzer-pro/requirements.txt
google-generativeai
python-dotenv
Pillow
pandas